requirements.txt
streamlit
PILLOW
urllib3
pdfplumber
beautifulsoup4
PyPDF2
google-auth
google-auth-oauthlib
google-api-python-client
jsonpickle
faiss-cpu
openai
duckduckgo-search
langchain